Remove json_dispatcher folder
Les fichiers .inc, on peut les mettre dans le dossier inc/ :)
